requirements.txt
PyYAML>=6.0.2
tiktoken>=0.10.0
anyio>=4.0.0
openai>=1.99.1
mcp>=1.12.3
google-genai>=1.28.0
ollama>=0.5.2

